NVDATool: Moved format_nested_section func to utils.py
