+++ site_scons/site_tools/NVDATool/utils.py
@@ -1,3 +1,8 @@
+from collections.abc import Mapping
+
+from .typings import Strable
+
+
 
 def _(arg: str) -> str:
 	"""
@@ -5,3 +10,15 @@ def _(arg: str) -> str:
 	Needed for recognizing strings for translation by Gettext.
 	"""
 	return arg
+
+
+def format_nested_section(
+	section_name: str,
+	data: Mapping[str, Mapping[str, Strable]]
+) -> str:
+	lines = [f"\n[{section_name}]"]
+	for item_name, inner_dict in data.items():
+		lines.append(f"[[{item_name}]]")
+		for key, val in inner_dict.items():
+			lines.append(f"{key} = {val}")
+	return "\n".join(lines) + "\n"
